Refactor: component 페이지별 폴더화 상혁씨 특정 페이지에서만 쓰이는 컴포넌트들은 커밋 참고하시고 변경 부탁드려요!
