+++ client/src/Pages/Login.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import LoginFooter from '../Components/LoginFooter'
+import LoginFooter from '../Components//Login/LoginFooter'
 import Id from '../Components/Id'
